v2-2024

+++ ui/src/components/bookings/BookingsList.tsx
@@ -7,18 +7,21 @@ import { Booking } from '../../types/bookings/Booking';
 import { modifyColumn } from '../../utils/modifyColumn';
 import { bookingsColumnDef } from './bookingsColumnDef';
 import { type MRT_ColumnDef } from 'mantine-react-table';
+import { removeEmptyKeys } from '../../utils/removeEmptyKeys';
+import { getChangedValues } from '../../utils/getChangedValues';
 
 const BookingList = () => {
 
   const dispatch = useAppDispatch();
   const { data, loading, error } = useAppSelector((state) => state.bookings)[0];
   const { currentPage } = useAppSelector((state) => state.navigation);
+  const { loggedUserId } = useAppSelector((state) => state.session);
+
   const initialState = {
     showColumnFilters: false,
     showGlobalFilter: true,
     columnVisibility: {
       id: false,
-      password: false,
     },
   }
 
@@ -26,24 +29,37 @@ const BookingList = () => {
     dispatch(fetchBookings())
   }, [dispatch]);
 
-  const handleCreate = async (dataCreate: Booking) => {
-    await dispatch(createBooking(dataCreate));
-    dispatch(fetchBookings());
+  const handleCreate = async (dataCreate: Partial<Booking>) => {
+    const dataCreated = removeEmptyKeys(dataCreate, loggedUserId);
+    if(Object.keys(dataCreated).some(key => key !== 'createdBy') && Object.keys(dataCreated).length !== 0){
+      await dispatch(createBooking(dataCreated));
+      dispatch(fetchBookings());
+    }
   };
 
-  const handleUpdate = async (dataUpdate: Booking) => {
-    const bookingId = dataUpdate.id;
-    const bookingData = dataUpdate;
-    await dispatch(updateBooking({ bookingId, bookingData }));
-    dispatch(fetchBookings());
+  const handleUpdate = async (originalData: Partial<Booking>, dataUpdated: Partial<Booking>) => {
+    const dataId = originalData.id;
+    const dataUpdate = getChangedValues(originalData, dataUpdated, loggedUserId);
+    if (typeof dataId === 'number') {
+      if(Object.keys(dataUpdate).some(key => key !== 'updatedBy') && Object.keys(dataUpdate).length !== 0){
+        await dispatch(updateBooking({ bookingId:dataId, bookingData:dataUpdate }));
+        dispatch(fetchBookings());
+      }
+    }else{
+      console.error('Booking ID is undefined.');
+    }
   };
 
-  const handleDelete = async (dataDelete: Booking, count: number, iterator: number) => {
-    await dispatch(deleteBooking(dataDelete.id));
-    if (count === iterator) { dispatch(fetchBookings()); }
+  const handleDelete = async (dataDelete: Partial<Booking>, count: number, iterator: number) => {
+    if (typeof dataDelete.id === 'number') {
+      await dispatch(deleteBooking(dataDelete.id));
+      if (count === iterator) { dispatch(fetchBookings()); }
+    }else{
+      console.error('Booking ID is undefined.');
+    }
   };
 
-  const modifiedColumns = useMemo<MRT_ColumnDef<Booking>[]>(() => modifyColumn(data[0]?.columns || [], bookingsColumnDef), [data]);
+  const modifiedColumns = useMemo<MRT_ColumnDef<Partial<Booking>>[]>(() => modifyColumn(data[0]?.columns || [], bookingsColumnDef), [data]);
 
   return (
     <Table
